- included <vw/Cartography/GeoReference.h> in Cartography/FileIO.h - added BConvex::num_facets() and BConvex::bounding_box() - added SpatialTree::closest() - made SpatialTree able to handle very small GeomPrimitives by adding a limit on how far down to create the tree when such a primitive is added - fixed a compiler warning in Math/Geometry.h

+++ src/vw/Math/Geometry.h
@@ -78,7 +78,7 @@ namespace math {
       }
     
       for (unsigned int i = 0; i < dimensions; ++i)
-        result[i] /= double(num_points);
+        result[i] /= num_points;
     
       return result;
     }
